Move rwlock package to pkg directory. Leave adapter package with alias types and forward calls. New rwlock contruction function accepts Options by value. Fixed typo in comments.
